pyproject.toml
[project]
name = "metalgate-code"
version = "0.1.0"
description = "ACP (Agent Communication Protocol) code agent featuring in-session dynamic deterministic tool skills and optional cross-session persistent memory."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents==0.5.5",
    "deepagents-acp==0.0.6",
    "deepagents-cli==0.0.46",
    "agent-client-protocol==0.9.0",
    "isort>=8.0.0",
    "langchain-community>=0.4.0",
    "langchain-mcp-adapters>=0.2.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.1.0",
    "mem0ai>=0.1.0",
    "chromadb>=0.5.0",
    "tiktoken>=0.12.0",
    "markdownify>=1.2.0",
    "pyyaml>=6.0.0",
    "tavily-python>=0.7.0",
    "tree-sitter",
    "tree-sitter-python",
    "jedi>=0.19.0",
    "sqlalchemy>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "ty",
    "harbor>=0.6.0"
]
